Extract operators into seperate folder and slight refactoring
